fix test result

+++ lib/classes/test_result.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 class TestResult {
   int speed;
-  int? jitter;
+  double? jitter;
   int? loss;
   TestResult({
     required this.speed,
